Add ClassroomJoinView to join classrooms
The users can click on `Join Class` on front page
and then select the classrooms they want to join
and they will be added to the required Classrooms.
Also instead of a class Student, I made use of the
existing UserProfile model to link "students" and
Classrooms. See accounts/models.py for the change made.